requirements.txt
django~=5.2.8
djangorestframework
pillow
django-filter
djangorestframework-simplejwt